place styling inline with components as they are
